gvim gtk config won't be rewrited after install.sh

+++ install.sh
@@ -18,7 +18,9 @@ fi
 if [ ! -f $INSTALL_DIR/.vimrc.user.after ]; then
 	cp vimrc.user.after $INSTALL_DIR/.vimrc.user.after
 fi
-cp gtkrc-2.0.mine $INSTALL_DIR/.gtkrc-2.0.mine
+if [ ! -f $INSTALL_DIR/.gtkrc-2.0.mine ]; then
+	cp gtkrc-2.0.mine $INSTALL_DIR/.gtkrc-2.0.mine
+fi
 
 echo
 echo "To enable fast search in ctrlp, please install ag"
